Fix date timezone, color tags in edit task form
