+++ src/shared/Form.js
@@ -4,6 +4,6 @@ export const Form = styled.form`
     display: flex;
     flex-direction: column;
     align-items: center;
-    gap: 32px;
+    gap: 24px;
     justify-content: space-between;
     padding-top: 16px;`
