+++ src/store/todos/thunk.js
@@ -1,3 +1,4 @@
+import moment from 'moment';
 import { api, baseURL } from '../../api'
 import { addTaskAction, deleteTaskAction, editTaskAction, saveTasksAction } from './actions'
 
@@ -16,7 +17,7 @@ export function fetchTasks() {
 }
 
 export function addTask({title, description, tags, deadline}) {
-    const newTask = {title, description, tags, deadline, isCompleted: false}
+    const newTask = {title, description, tags, deadline, isCompleted: false, createdAt: moment(new Date()).format()}
     return (dispatch, getState) => {
         baseURL.post(api.tasks(), newTask)
             .then(res => {
